+++ js/constants.js
@@ -1,0 +1,9 @@
+const NAMES = [
+  "Sofia", "Egor", "Oleg", "Alexandra", "Semion", "Valentina", "Ksenia", "Vadim", "Vika", "Kirill",
+  "Denis", "Alexandr", "Nikita", "Nastya", "Sergey", "Tanya", "Sveta", "Olga", "Veronika", "Yana",
+  "Lena", "Zhenya", "Timur", "Svyatozar", "Fillimon", "David", "Murat", "Leonid", "Masha", "Vera"
+];
+
+const TEXT_FOR_MESSAGE = "Всё отлично! В целом всё неплохо. Но не всё. Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально. Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше. Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше. Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!";
+
+export {NAMES, TEXT_FOR_MESSAGE};
